+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/requests/MarkUnreadRequest.kt
@@ -21,4 +21,5 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 internal data class MarkUnreadRequest(
     val message_id: String,
+    val thread_id: String? = null,
 )